+++ .eslintrc.cjs
@@ -48,7 +48,7 @@ module.exports = {
   settings: {
     react: {
       pragma: 'React',
-      version: 'detect'
-    }
-  }
+      version: 'detect',
+    },
+  },
 };
