+++ src/app/countriesSlice.ts
@@ -19,5 +19,4 @@ export const countrySlice = createSlice({
   },
 });
 
-// export const { setCountry } = countrySlice.actions;
 export default countrySlice.reducer;
