+++ src/components/form-uncontrolled/FormUncontrolled.css
@@ -8,6 +8,7 @@
 .form-uncontrolled {
   display: flex;
   flex-direction: column;
+  background-color: #f0f8ff;
   width: 100%;
   gap: 10px;
 }
@@ -27,10 +28,13 @@
   width: 40%;
 }
 
+.input-container_height {
+  height: 60px;
+}
+
 .field__input {
   display: flex;
   align-items: center;
-  /* width: 40%; */
   font-size: 1.3rem;
   border-radius: 3px;
 }
@@ -58,6 +62,7 @@
 }
 
 .input__validate {
+  height: 30px;
   font-size: 0.8rem;
   font-weight: 700;
   color: #a50000;
