--- src/shared/RenderRouteOutlet.tsx
@@ -1,22 +1,0 @@
-import { ReactNode } from 'react';
-import { MemoryRouter, Routes, Route, Outlet } from 'react-router-dom';
-
-interface RenderRouteWithOutletContextProps<T> {
-  context: T;
-  children: ReactNode;
-}
-
-export const RenderRouteWithOutletContext = <T,>({
-  context,
-  children,
-}: RenderRouteWithOutletContextProps<T>) => {
-  return (
-    <MemoryRouter>
-      <Routes>
-        <Route path="/" element={<Outlet context={context as T} />}>
-          <Route index element={children} />
-        </Route>
-      </Routes>
-    </MemoryRouter>
-  );
-};
