--- src/shared/navigate.ts
@@ -1,11 +1,0 @@
-import { NavigateFunction } from 'react-router-dom';
-
-export default function navigateToPage(
-  navigate: NavigateFunction,
-  inputValue: string | undefined,
-  pageNumber: number
-) {
-  inputValue
-    ? navigate(`/search=${inputValue}&page=${pageNumber}`)
-    : navigate(`/page=${pageNumber}`);
-}
